As the number of mailboxes increases, the execution time of the `update_sogo_static_view` function will also increase. When a new mailbox is added, the information is saved in the mailbox table and is reflected in the `sogo_view`. The `update_sogo_static_view` function then updates the `_sogo_static_view` table accordingly.

The bottleneck is caused by this query:
```
REPLACE INTO _sogo_static_view (`c_uid`, `domain`, `c_name`, `c_password`, `c_cn`, `mail`, `aliases`, `ad_aliases`, `ext_acl`, `kind`, `multiple_bookings`)
      SELECT `c_uid`, `domain`, `c_name`, `c_password`, `c_cn`, `mail`, `aliases`, `ad_aliases`, `ext_acl`, `kind`, `multiple_bookings` from sogo_view
```

After some experimentation, I found a solution that doesn't suffer from performance issues (in my tests). When a mailbox is added, edited, or deleted, the `update_sogo_static_view` function will only try to update that specific mailbox in the `_sogo_static_view`. Furthermore, the mailbox table is used directly and the `sogo_view` view is skipped, as it is not possible otherwise.


@mkuron what do you think?
The issue was first mentioned here: https://github.com/mailcow/mailcow-dockerized/issues/4980. At that point, I only added a skip if SOGo was not being used.


Benchmark:
before
```
Created 0 accounts; the last 100 accounts took 0.00 seconds
Created 100 accounts; the last 100 accounts took 9.61 seconds
Created 200 accounts; the last 100 accounts took 9.86 seconds
Created 300 accounts; the last 100 accounts took 10.47 seconds
Created 400 accounts; the last 100 accounts took 11.82 seconds
```

after
```
Created 0 accounts; the last 100 accounts took 0.00 seconds
Created 100 accounts; the last 100 accounts took 9.37 seconds
Created 200 accounts; the last 100 accounts took 9.72 seconds
Created 300 accounts; the last 100 accounts took 9.57 seconds
Created 400 accounts; the last 100 accounts took 9.60 seconds
Created 500 accounts; the last 100 accounts took 9.78 seconds
Created 600 accounts; the last 100 accounts took 9.89 seconds
Created 700 accounts; the last 100 accounts took 10.08 seconds
Created 800 accounts; the last 100 accounts took 10.62 seconds
Created 900 accounts; the last 100 accounts took 9.91 seconds
Created 1000 accounts; the last 100 accounts took 10.36 seconds
Created 1100 accounts; the last 100 accounts took 10.82 seconds
```