@mkuron 2023-04-15
Yes, that looks like a good idea.